requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
